# src/main/java/com/mishchuk/onlineschool/controller/dto/ModuleCreateDto.java
package com.mishchuk.onlineschool.controller.dto;

public record ModuleCreateDto(
        String name,
        java.util.UUID courseId,
        String description,
        java.util.List<java.util.UUID> lessonIds) {
}