# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Online School API
  description: REST API для платформи онлайн-навчання. Включає управління курсами, модулями, уроками, користувачами, реєстрацією на курси та файлами.
  version: 1.0.0
  contact:
    name: API Support
    email: support@onlineschool.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /modules:
    get:
      tags:
        - Modules
      summary: Отримати всі модулі
      operationId: getAllModules
      responses:
        '200':
          description: Список модулів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleDto'
        '204':
          description: Немає модулів
    post:
      tags:
        - Modules
      summary: Створити новий модуль
      operationId: createModule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleCreateDto'
      responses:
        '201':
          description: Модуль створено
  /modules/{id}:
    get:
      tags:
        - Modules
      summary: Отримати модуль за ID
      operationId: getModule
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Деталі модуля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '404':
          description: Модуль не знайдено
    put:
      tags:
        - Modules
      summary: Оновити модуль
      operationId: updateModule
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpdateDto'
      responses:
        '204':
          description: Модуль оновлено
        '404':
          description: Модуль не знайдено
    delete:
      tags:
        - Modules
      summary: Видалити модуль
      operationId: deleteModule
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Модуль видалено

  # ==================== Lessons ====================
  /lessons:
    post:
      tags:
        - Lessons
      summary: Створити новий урок
      operationId: createLesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonCreateDto'
      responses:
        '201':
          description: Урок створено
  /lessons/{id}:
    get:
      tags:
        - Lessons
      summary: Отримати урок за ID
      operationId: getLesson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Деталі уроку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonDto'
        '404':
          description: Урок не знайдено
    put:
      tags:
        - Lessons
      summary: Оновити урок
      operationId: updateLesson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonUpdateDto'
      responses:
        '204':
          description: Урок оновлено
        '404':
          description: Урок не знайдено
    delete:
      tags:
        - Lessons
      summary: Видалити урок
      operationId: deleteLesson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Урок видалено

  # ==================== Enrollments ====================
...
  /persons:
    get:
      tags:
        - Persons
      summary: Отримати всіх користувачів
      operationId: getAllPersons
      responses:
        '200':
          description: Список користувачів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonDto'
        '204':
          description: Немає користувачів
    post:
      tags:
        - Persons
      summary: Створити користувача
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreateDto'
      responses:
        '201':
          description: Користувача створено
  /persons/{id}:
    get:
      tags:
        - Persons
      summary: Отримати користувача за ID
      operationId: getPerson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Деталі користувача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDto'
        '404':
          description: Користувача не знайдено
    put:
      tags:
        - Persons
      summary: Оновити користувача
      operationId: updatePerson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonUpdateDto'
      responses:
        '204':
          description: Користувача оновлено
        '404':
          description: Користувача не знайдено
    delete:
      tags:
        - Persons
      summary: Видалити користувача
      operationId: deletePerson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Користувача видалено



  # ==================== Files ====================
  /api/files/upload:
    post:
      tags:
        - Files
      summary: Завантажити файл
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл для завантаження (PDF, JPEG, PNG)
                entityType:
                  type: string
                  description: Тип entity (COURSE, MODULE, LESSON)
                  example: COURSE
                entityId:
                  type: string
                  format: uuid
                  description: ID entity, до якої прив'язаний файл
      responses:
        '200':
          description: Файл завантажено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '400':
          description: Невалідний файл
  /api/files/{fileId}:
    get:
      tags:
        - Files
      summary: Завантажити файл
      operationId: downloadFile
      security: []
      parameters:
        - in: path
          name: fileId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Файл
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'attachment; filename="..."'
        '404':
          description: Файл не знайдено
    delete:
      tags:
        - Files
      summary: Видалити файл
      operationId: deleteFile
      parameters:
        - in: path
          name: fileId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Файл видалено
        '404':
          description: Файл не знайдено
  /api/files/entity/{entityType}/{entityId}:
    get:
      tags:
        - Files
      summary: Отримати файли для entity
      operationId: getFilesForEntity
      parameters:
        - in: path
          name: entityType
          required: true
          schema:
            type: string
            enum: [COURSE, MODULE, LESSON]
        - in: path
          name: entityId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список файлів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: UUID ідентифікатор
  schemas:
...
    PersonCreateDto:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 6
    PersonDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, BLOCKED]
    PersonUpdateDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email



    # ==================== Course Schemas ====================
...
    ModuleCreateDto:
      type: object
      required:
        - name
        - description
        - courseId
      properties:
        name:
          type: string
        description:
          type: string
        courseId:
          type: string
          format: uuid
    ModuleDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        courseId:
          type: string
          format: uuid
        courseName:
          type: string
        orderIndex:
          type: integer
    ModuleUpdateDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        orderIndex:
          type: integer

    # ==================== Lesson Schemas ====================
    LessonCreateDto:
      type: object
      required:
        - title
        - content
        - moduleId
      properties:
        title:
          type: string
        content:
          type: string
        moduleId:
          type: string
          format: uuid
        orderIndex:
          type: integer
    LessonDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        moduleId:
          type: string
          format: uuid
        moduleName:
          type: string
        orderIndex:
          type: integer
        duration:
          type: integer
          description: Тривалість в хвилинах
    LessonUpdateDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        orderIndex:
          type: integer
        duration:
          type: integer

    # ==================== Enrollment Schemas ====================
...
    FileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        originalName:
          type: string
        contentType:
          type: string
          example: application/pdf
        fileSize:
          type: integer
          format: int64
          description: Розмір в байтах
        uploadedAt:
          type: string
          format: date-time
        downloadUrl:
          type: string
          example: /api/files/uuid
        relatedEntityType:
          type: string
          enum: [COURSE, MODULE, LESSON]
        relatedEntityId:
          type: string
          format: uuid